+++ src/design/plone/contenttypes/indexers/common.py
@@ -1,11 +1,9 @@
 # -*- coding: utf-8 -*-
 from plone.indexer.decorator import indexer
-from plone.dexterity.interfaces import IDexterityContainer
-from plone.dexterity.interfaces import IDexterityItem
+from plone.dexterity.interfaces import IDexterityContent
 
 
-@indexer(IDexterityContainer)
-@indexer(IDexterityItem)
+@indexer(IDexterityContent)
 def tassonomia_argomenti(context, **kw):
     return [
         x.to_object.Title()
@@ -14,15 +12,13 @@ def tassonomia_argomenti(context, **kw):
     ]
 
 
-@indexer(IDexterityContainer)
-@indexer(IDexterityItem)
+@indexer(IDexterityContent)
 def ufficio_responsabile(context, **kw):
     uffici = getattr(context.aq_base, "ufficio_responsabile", [])
     return [ufficio.UID() for ufficio in filter(bool, [x.to_object for x in uffici])]
 
 
-@indexer(IDexterityContainer)
-@indexer(IDexterityItem)
+@indexer(IDexterityContent)
 def parent(context):
     obj_parent = context.aq_parent
     return {